+++ constants/constants.go
@@ -1,8 +1,8 @@
 package constants
 
 const (
-	DateLayout       = "20060102"
-	DefaultPort      = "7540"
-	DefaultDbName    = "scheduler.db"
+	DateLayout  = "20060102"
+	DefaultPort = "7540"
+
 	TasksNumberLimit = 50
 )
