+++ handlers/done_task.go
@@ -5,7 +5,6 @@ import (
 	"go_final_project/db"
 	"go_final_project/models"
 	"go_final_project/services"
-	"go_final_project/utils"
 	"net/http"
 	"strconv"
 	"time"
@@ -16,35 +15,35 @@ func HandleTaskDone(w http.ResponseWriter, req *http.Request) {
 	id, err := strconv.Atoi(req.URL.Query().Get("id"))
 
 	if err != nil {
-		utils.HandleError(w, http.StatusBadRequest, errors.New("Неверный идентификатор"))
+		HandleError(w, http.StatusBadRequest, errors.New("Неверный идентификатор"))
 		return
 	}
 	task, err = db.GetTaskById(id)
 	if err != nil {
-		utils.HandleError(w, http.StatusBadRequest, err)
+		HandleError(w, http.StatusBadRequest, err)
 		return
 	}
 	if task.Repeat == "" {
 		err = db.DeleteById(id)
 		if err != nil {
-			utils.HandleError(w, http.StatusInternalServerError, err)
+			HandleError(w, http.StatusInternalServerError, err)
 			return
 		}
-		utils.WriteNormalResponse(w, "")
+		HandleNormalResponse(w, "")
 		return
 	}
 	now := time.Now().Truncate(24 * time.Hour)
 	task.Date, err = services.NextDate(now, task.Date, task.Repeat)
 	if err != nil {
-		utils.HandleError(w, http.StatusInternalServerError, err)
+		HandleError(w, http.StatusInternalServerError, err)
 		return
 	}
 	err = db.PutTask(task)
 	if err != nil {
-		utils.HandleError(w, http.StatusInternalServerError, err)
+		HandleError(w, http.StatusInternalServerError, err)
 		return
 	}
 
-	utils.WriteNormalResponse(w, "")
+	HandleNormalResponse(w, "")
 	return
 }
